+++ src/api/controllers/setting/settings.controller.ts
@@ -3,6 +3,7 @@ import constants from '../../../utils/constants.util';
 import responseHelper from '../../../utils/responseHelper.util';
 import CreditorService from '../../services/creditor.service';
 import SettingsService from '../../services/settings.service';
+import commonUtil from '../../../utils/common.util';
 
 class SettingsController {
   protected settingsService: SettingsService;
@@ -12,7 +13,23 @@ class SettingsController {
   }
   addSettings = async (req: Request, res: Response) => {
     try {
-      const response = await this.settingsService.addSettings(req);
+      let keyword = '';
+      if (String(req.query.type) === 'template') {
+        keyword = 'addNotificationTemplate';
+        const checkPermission = await commonUtil.checkPermission(keyword, req);
+        if (!checkPermission)
+          return res
+            .status(constants.CODE.BAD_REQUEST)
+            .send(
+              responseHelper.get4xxResponse(
+                'You do not have permission to perform this operation'
+              )
+            );
+      }
+      if (String(req.query.type) === 'payments') {
+        keyword = 'editPaymentsNotificationSettings';
+      }
+      const response = await this.settingsService.addSettings(req, keyword);
       if (!response[0]) {
         return res
           .status(constants.CODE.BAD_REQUEST)
@@ -26,7 +43,7 @@ class SettingsController {
         })
       );
     } catch (error) {
-      console.log(error.message);
+      console.log(error);
       return res
         .status(constants.CODE.BAD_REQUEST)
         .send(responseHelper.get4xxResponse(constants.Messages.EXCEPTION));
@@ -35,7 +52,23 @@ class SettingsController {
 
   getSettings = async (req: Request, res: Response) => {
     try {
-      const response = await this.settingsService.getSettings();
+      const templatePermission = await commonUtil.checkPermission(
+        'viewNotificationTemplates',
+        req
+      );
+      const paymentsPermission = await commonUtil.checkPermission(
+        'viewPaymentsNotificationSettings',
+        req
+      );
+      const customFieldsPermission = await commonUtil.checkPermission(
+        'viewCustomFields',
+        req
+      );
+      const response = await this.settingsService.getSettings(
+        templatePermission,
+        paymentsPermission,
+        customFieldsPermission
+      );
       return res.status(constants.CODE.OK).send(
         responseHelper.get2xxResponse({
           statusCode: constants.CODE.OK,
@@ -53,6 +86,18 @@ class SettingsController {
 
   addCustomField = async (req: Request, res: Response) => {
     try {
+      const checkPermission = await commonUtil.checkPermission(
+        'addCustomFields',
+        req
+      );
+      if (!checkPermission)
+        return res
+          .status(constants.CODE.BAD_REQUEST)
+          .send(
+            responseHelper.get4xxResponse(
+              'You do not have permission to perform this operation'
+            )
+          );
       const response = await this.settingsService.addCustomField(req);
       if (!response[0]) {
         return res
@@ -75,6 +120,18 @@ class SettingsController {
   };
   editCustomField = async (req: Request, res: Response) => {
     try {
+      const checkPermission = await commonUtil.checkPermission(
+        'editCustomFields',
+        req
+      );
+      if (!checkPermission)
+        return res
+          .status(constants.CODE.BAD_REQUEST)
+          .send(
+            responseHelper.get4xxResponse(
+              'You do not have permission to perform this operation'
+            )
+          );
       const response = await this.settingsService.editCustomField(req);
       if (!response[0]) {
         return res
@@ -187,6 +244,18 @@ class SettingsController {
   };
   deleteCustomField = async (req: Request, res: Response) => {
     try {
+      const checkPermission = await commonUtil.checkPermission(
+        'deleteCustomFields',
+        req
+      );
+      if (!checkPermission)
+        return res
+          .status(constants.CODE.BAD_REQUEST)
+          .send(
+            responseHelper.get4xxResponse(
+              'You do not have permission to perform this operation'
+            )
+          );
       const response = await this.settingsService.deleteCustomField(req);
       if (!response[0]) {
         return res
@@ -209,6 +278,18 @@ class SettingsController {
 
   editNotificationTemplate = async (req: Request, res: Response) => {
     try {
+      const checkPermission = await commonUtil.checkPermission(
+        'editNotificationTemplate',
+        req
+      );
+      if (!checkPermission)
+        return res
+          .status(constants.CODE.BAD_REQUEST)
+          .send(
+            responseHelper.get4xxResponse(
+              'You do not have permission to perform this operation'
+            )
+          );
       const response = await this.settingsService.editNotificationTemplate(req);
       if (!response[0]) {
         return res
@@ -231,6 +312,18 @@ class SettingsController {
 
   deleteNotificationTemplate = async (req: Request, res: Response) => {
     try {
+      const checkPermission = await commonUtil.checkPermission(
+        'deleteNotificationTemplate',
+        req
+      );
+      if (!checkPermission)
+        return res
+          .status(constants.CODE.BAD_REQUEST)
+          .send(
+            responseHelper.get4xxResponse(
+              'You do not have permission to perform this operation'
+            )
+          );
       const response =
         await this.settingsService.deleteNotificationTemplate(req);
       if (!response[0]) {
